[Modify] Refactor filter class :sun_with_face:
